Fikset logoerne i header. Ændret userId fra int til long, så den kunne kører

+++ src/main/java/com/example/oenskeseddel/services/WishlistService.java
@@ -17,7 +17,7 @@ public class WishlistService {
         this.userRepository = userRepository;
     }
 
-    public Wishlist createWishlist(Integer userId, String name) {
+    public Wishlist createWishlist(long userId, String name) {
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new RuntimeException("Bruger ikke fundet"));
